% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{colorsafe_palette}
\alias{colorsafe_palette}
\alias{discrete_palette}
\alias{diverging_palette}
\alias{qualitative_palette}
\alias{sequential_palette}
\alias{palettes}
\title{Colorblind-safe color palettes}
\usage{
colorsafe_palette(
  palette,
  n,
  chroma = c("high", "low", "variable"),
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

discrete_palette(
  palette,
  chroma = c("high", "low", "variable"),
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

diverging_palette(
  hue_pair = "Red-Blue",
  chroma = "variable",
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

qualitative_palette(
  name = "Set 1.5",
  chroma = c("high", "low", "variable"),
  grayscale_safe = FALSE,
  reverse = FALSE
)

sequential_palette(
  hue = "Red",
  chroma = c("high", "low", "variable"),
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

palettes(
  n,
  type = c("qualitative", "sequential", "diverging"),
  chroma = c("high", "low", "variable"),
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)
}
\arguments{
\item{palette}{Name of palette. Run \code{palettes()} to see palette names and hex codes.}

\item{n}{Number of colors to return. \code{n} may not be greater than the maximum
number colors for a given palette and chroma.}

\item{chroma}{Color chroma based on HCL colorspace. Any of \code{"high"}, \code{"low"},
or \code{"variable"}.}

\item{grayscale_safe}{If \code{TRUE}, palette must be grayscale safe.}

\item{reverse}{Reverse the order of the color palette.}

\item{luminance_range}{Minimum and maximum luminance values in HCL colorspace.}

\item{max_chroma}{Maximum chroma in HCL colorspace. Applicable when
\code{chroma = "variable"}.}

\item{hue_pair}{Name of a diverging palette.}

\item{name}{Name of a qualitative palette.}

\item{hue}{Name of a seqential palette.}

\item{type}{Type of palette. Must be \code{"qualitative"},  \code{"sequential"}, or
\code{"diverging"}.}
}
\value{
A color palette function, palette, or list of palettes (see description).
}
\description{
\code{colorsafe_palette} returns the named color palette.

\code{discrete_palette} returns a palette function for discrete scales.

\code{diverging_palette} returns a two-color diverging palette function.

\code{qualitative_palette} returns a multi-hue qualitative palette function.

\code{sequential_palette} returns a single-hue sequential palette function.

\code{palettes} returns a list of available palettes.
}
\examples{
pal <- colorsafe_palette("Paired 6")
colorspace::swatchplot(pal)

pal <- diverging_palette()(5)
colorspace::swatchplot(pal)
pal <- diverging_palette(reverse = TRUE, luminance_range = c(50,97))(5)
colorspace::swatchplot(pal)
pal <- diverging_palette("Chartreuse-Magenta", luminance_range = c(40,90))(6)
colorspace::swatchplot(pal, cvd = c("deutan","protan"))

pal <- qualitative_palette()(5)
colorspace::swatchplot(pal)
pal <- qualitative_palette("Set 2.5")(5)
colorspace::swatchplot(pal)
pal <- qualitative_palette(chroma = "low", grayscale_safe = TRUE)(5)
colorspace::swatchplot(pal, cvd = c("deutan","protan","desaturate"))

pal <- sequential_palette()(4)
colorspace::swatchplot(pal)
pal <- sequential_palette(chroma = "variable", luminance_range = c(50,97))(4)
colorspace::swatchplot(pal)
pal <- sequential_palette(chroma = "low")(4)
colorspace::swatchplot(pal)

names(palettes(type = "diverging", chroma = "variable"))
palettes(5, type = "qualitative", grayscale_safe = TRUE)
}
