% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colorsafe_continuous}
\alias{scale_colorsafe_continuous}
\alias{scale_color_colorsafe_continuous}
\alias{scale_fill_colorsafe_continuous}
\alias{scale_colorsafe_discrete}
\alias{scale_colour_colorsafe_discrete}
\alias{scale_fill_colorsafe_discrete}
\alias{scale_colorsafe_binned}
\alias{scale_color_colorsafe_binned}
\alias{scale_fill_colorsafe_binned}
\title{Colorblind-safe color scales}
\usage{
scale_colorsafe_continuous(
  ...,
  aesthetic = "colour",
  palette = "Red",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_color_colorsafe_continuous(
  ...,
  palette = "Red",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_fill_colorsafe_continuous(
  ...,
  palette = "Red",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_colorsafe_discrete(
  ...,
  aesthetic = "colour",
  palette = "Set 1.6",
  chroma = "low",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_colour_colorsafe_discrete(
  ...,
  palette = "Set 1.6",
  chroma = "low",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_fill_colorsafe_discrete(
  ...,
  palette = "Set 1.6",
  chroma = "low",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_colorsafe_binned(
  ...,
  aesthetic = "colour",
  palette = "Red-Blue",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_color_colorsafe_binned(
  ...,
  palette = "Red-Blue",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)

scale_fill_colorsafe_binned(
  ...,
  palette = "Red-Blue",
  chroma = "variable",
  grayscale_safe = FALSE,
  reverse = FALSE,
  luminance_range = c(20, 90),
  max_chroma = 105
)
}
\arguments{
\item{...}{Other arguments passed \code{ggplot2} functions \code{discrete_scale()},
\code{continuous_scale()}, or \code{binned_scale()}.}

\item{aesthetic}{\code{"colour"} or \code{"fill"} aesthetic.}

\item{palette}{Name of palette. Run \code{\link{palettes}()} to see palette names and hex
codes.}

\item{chroma}{Color chroma based on HCL colorspace. Any of \code{"high"}, \code{"low"},
or \code{"variable"}.}

\item{grayscale_safe}{If \code{TRUE}, palette must be grayscale safe.}

\item{reverse}{Reverse the order of the color palette.}

\item{luminance_range}{Minimum and maximum luminance values in HCL colorspace.}

\item{max_chroma}{Maximum chroma in HCL colorspace. Applicable when
\code{chroma = "variable"}.}
}
\description{
The \code{colorsafe} scales are colorblind safe and most palettes safely
convert to grayscale.
}
\examples{
p <- ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
       geom_tile()
p + scale_fill_colorsafe_binned()
p + scale_fill_colorsafe_continuous(reverse = TRUE)

p <- ggplot(mtcars, aes(mpg, wt)) +
       geom_point(aes(colour = factor(cyl)), size = 2)
p + scale_colorsafe_discrete(palette = "Set 1.4", chroma = "high")
}
